stage0-module3-taskFinal v2.0

+++ src/main/java/lang/print/gaps/finalModuleTask/TemperatureConverter.java
@@ -3,7 +3,7 @@ package lang.print.gaps.finalModuleTask;
 public class TemperatureConverter {
     public void toFahrenheit(int temperatureCelsius){
         // (30 °C × 9/5) + 32 = 86 °F
-        int fahrenheit = (temperatureCelsius * 9/5) + 32;
+        double fahrenheit = (temperatureCelsius * 9.0/5.0) + 32.0;
         System.out.println(fahrenheit);
     }
 }
